requirements.txt
mysql-connector
tqdm